# server/docs/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "tubu.io API Documentation - ",
    "version": "0.2.0",
    "title": "TUBU.IO API",
    "contact": {
      "email": "dev@tubu.io"
    }
  },
...
  "paths": {
...
    "/api/market/{id}": {
      "post": {
        "tags": ["Market"],
        "summary": "Deploy contracts from marketplace",
        "description": "",
        "operationId": "deployMarketContract",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "network_id": { "type": "integer", "required": true },
                  "application_id": { "type": "integer", "required": true },
                  "name": { "type": "string", "required": true },
                  "description": { "type": "string", "required": true },
                  "tag": { "type": "string", "required": true },
                  "args": { "type": "string|array", "required": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "abi": {
                      "type": "string"
                    },
                    "bytecode": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "string"
                    },
                    "market_version": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      },
      "get": {
        "tags": ["Market"],
        "summary": "Get the public contracts and the user's own contracts from marketplace",
        "description": "",
        "operationId": "getMarketContract",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "updated_at": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "totalCount": {
                          "type": "integer"
                        },
                        "totalPage": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      }
    },
...
  },
  "security": [
    {
      "UserToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "UserToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
...
    "parameters": {
...
      "id": {
        "in": "path",
        "name": "id",
        "description": "ID for contract",
        "schema": {
          "minimum": 1,
          "default": 1,
          "type": "integer"
        }
      },
...
    }
  }
}